feat: implement rate limiting middleware across Supabase Edge Functions
